requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
croniter==3.0.3
discord-typings==0.9.0
distro==1.9.0
emoji==2.14.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
multidict==6.1.0
openai==1.52.0
propcache==0.2.0
py-cord==2.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
tomli==2.0.2
tqdm==4.66.5
typing_extensions==4.12.2
yarl==1.15.5
